+++ instrumentation/opentelemetry-instrumentation-logging/src/opentelemetry/instrumentation/logging/__init__.py
@@ -55,6 +55,10 @@ class LoggingInstrumentor(BaseInstrumentor):  # pylint: disable=empty-docstring
 
         {DEFAULT_LOGGING_FORMAT}
 
+        def log_hook(span: Span, record: LogRecord):
+                if span and span.is_recording():
+                    record.custom_user_attribute_from_log_hook = "some-value"
+
     Args:
         tracer_provider: Tracer provider instance that can be used to fetch a tracer.
         set_logging_format: When set to True, it calls logging.basicConfig() and sets a logging format.
@@ -66,11 +70,13 @@ class LoggingInstrumentor(BaseInstrumentor):  # pylint: disable=empty-docstring
             logging.WARN
             logging.ERROR
             logging.FATAL
+        log_hook: execute custom logic when record is created
 
     See `BaseInstrumentor`
     """
 
     _old_factory = None
+    _log_hook = None
 
     def instrumentation_dependencies(self) -> Collection[str]:
         return _instruments
@@ -80,6 +86,7 @@ class LoggingInstrumentor(BaseInstrumentor):  # pylint: disable=empty-docstring
         provider = kwargs.get("tracer_provider", None) or get_tracer_provider()
         old_factory = logging.getLogRecordFactory()
         LoggingInstrumentor._old_factory = old_factory
+        LoggingInstrumentor._log_hook = kwargs.get("log_hook", None)
 
         service_name = None
 
@@ -107,6 +114,14 @@ class LoggingInstrumentor(BaseInstrumentor):  # pylint: disable=empty-docstring
                 if ctx != INVALID_SPAN_CONTEXT:
                     record.otelSpanID = format(ctx.span_id, "016x")
                     record.otelTraceID = format(ctx.trace_id, "032x")
+                    if callable(LoggingInstrumentor._log_hook):
+                        try:
+                            LoggingInstrumentor._log_hook(  # pylint: disable=E1102
+                                span, record
+                            )
+                        except Exception:  # pylint: disable=W0703
+                            pass
+
             return record
 
         logging.setLogRecordFactory(record_factory)