+++ instrumentation/opentelemetry-instrumentation-logging/tests/test_logging.py
@@ -64,6 +64,10 @@ class TestLoggingInstrumentorProxyTracerProvider(TestBase):
             self.assertEqual(record.otelServiceName, "")
 
 
+def log_hook(span, record):
+    record.custom_user_attribute_from_log_hook = "some-value"
+
+
 class TestLoggingInstrumentor(TestBase):
     @pytest.fixture(autouse=True)
     def inject_fixtures(self, caplog):
@@ -149,6 +153,27 @@ class TestLoggingInstrumentor(TestBase):
         basic_config_mock.assert_called_with(
             format="%(message)s span_id=%(otelSpanID)s", level=logging.WARNING
         )
+
+    def test_log_hook(self):
+        LoggingInstrumentor().uninstrument()
+        LoggingInstrumentor().instrument(
+            set_logging_format=True,
+            log_hook=log_hook,
+        )
+        with self.tracer.start_as_current_span("s1") as span:
+            span_id = format(span.get_span_context().span_id, "016x")
+            trace_id = format(span.get_span_context().trace_id, "032x")
+            with self.caplog.at_level(level=logging.INFO):
+                logger = logging.getLogger("test logger")
+                logger.info("hello")
+                self.assertEqual(len(self.caplog.records), 1)
+                record = self.caplog.records[0]
+                self.assertEqual(record.otelSpanID, span_id)
+                self.assertEqual(record.otelTraceID, trace_id)
+                self.assertEqual(record.otelServiceName, "unknown_service")
+                self.assertEqual(
+                    record.custom_user_attribute_from_log_hook, "some-value"
+                )
 
     def test_uninstrumented(self):
         with self.tracer.start_as_current_span("s1") as span: